Optional SQL usage for `add_atmo_dep()`
This change allows the user more control over the behavior of `add_atmo_dep()`. The function can now either update the SQL and write the atmodep.cli file, or only write the atmodep.cli file.

Additionally, the location of the SQL database and location of where the text file is written can now be different.